requirements.txt
APScheduler==3.3.1
beautifulsoup4==4.5.1
colorama==0.3.7
coloredlogs==5.2
configparser==3.5.0
decorator==4.0.10
enum34==1.1.6
eyeD3==0.7.11
flake8==3.2.1
flake8-import-order==0.11
funcsigs==1.0.2
future==0.16.0
gmusicapi==10.1.0
gpsoauth==0.4.0
httplib2==0.9.2
humanfriendly==2.1
logutils==0.3.3
mccabe==0.5.3
MechanicalSoup==0.6.0
mock==2.0.0
mutagen==1.35.1
oauth2client==4.0.0
pbr==1.10.0
pep8==1.5.7
proboscis==1.2.6.0
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycodestyle==2.2.0
pycryptodomex==3.4.3
pyflakes==0.8.1
python-dateutil==2.6.0
rainbow-logging-handler==2.2.2
requests==2.11.1
rsa==3.4.2
slackbot==0.4.1
slacker==0.9.24
validictory==1.1.0
websocket-client==0.37.0
youtube-dl==2017.1.10
google-api-python-client==1.6.1
